requirements.txt
selenium
webdriver-manager
pyautogui
websocket-client

